+++ test/routes/productsRouter.test.js
@@ -19,7 +19,10 @@ describe("[Products router] tests", () => {
   before(async function () {
     this.timeout(10000);
     await mongoose.connect(MONGO_URL);
-    await mongoose.connection.collections["products"].drop();
+    if(mongoose.connection.collections["sessions"])
+    await mongoose.connection.collections["sessions"].drop()
+    if(mongoose.connection.collections["products"])
+      await mongoose.connection.collections["products"].drop();
     cookies = await getPremiumMockUserCookiesForTest(requester);
   });
